View model approach is working. I Have not tested the onAppear approach given here but it should work. HJopefully this isn't too confusing, My initial advice wasn't quite correct, and i ended up putting together an example of a view model for this, since it works better than the onAppear approach.
Because the view model is a bit of a major change in structure i also left in commented out code that you could enable to try the .onAppear approach. I find that there are UI glitches with the onAppear approach. For example, in a page with a segmented picker control, if you change the value onAppear you get a weird 'wobble' in the picker control when the view appears as the value is set. This doesn't happen with the view model approach.

I also noted a couple of things that might help regarding layout stuff. Feel free to ignore if im misunderstanding your design goal!